Remove dependency to pg_worker_list.conf, add a new UDF
This commit removes the dependency to pg_worker_list.conf, and instead
adds a new UDF which can add nodes to the cluster, namely cluster_add_node().

The worker nodes are hold on a new metadata table, namely pg_dist_node.

Before this change, worker node cache was hold on a shared memory hash table.
Since it's not possible anymore, we moved the cache to the per-session cache
similar to pg_dist_partition or pg_shard_shard. We also utilized the same cache
invalidation approach.